+++ Core/include/Acts/Definitions/Units.hpp
@@ -52,16 +52,17 @@ namespace Acts {
 ///
 /// .. code-block:: cpp
 ///
-///    double length = 1 * Acts::UnitConstants::m;       // length == 1000.0
-///    double momentum = 100 * Acts::UnitConstants::MeV; // momentum == 0.1
+///    long double length = 1 * Acts::UnitConstants::m;       // length ==
+///    1000.0 long double momentum = 100 * Acts::UnitConstants::MeV; // momentum
+///    == 0.1
 ///
 /// The conversion of a value in native units into the equivalent value in a
 /// specific other unit is computed by dividing with the relevant unit, e.g.
 ///
 /// .. code-block:: cpp
 ///
-///    double length = 10.0;                               // native units mm
-///    double lengthInM = length / Acts::UnitConstants::m; // == 0.01;
+///    long double length = 10.0;                               // native units
+///    mm long double lengthInM = length / Acts::UnitConstants::m; // == 0.01;
 ///
 /// To further simplify the usage, physical quantities can also be expressed via
 /// `C++ user literals
@@ -73,10 +74,10 @@ namespace Acts {
 ///
 ///    using namespace Acts::UnitLiterals;
 ///
-///    double length = 1_m;                     // == 1000.0
-///    double momentum = 1.25_TeV;              // == 1250.0
-///    double lengthInUm = length / 1_um;       // == 1000000.0
-///    double momentumInMeV = momentum / 1_MeV; // == 1250000.0
+///    long double length = 1_m;                     // == 1000.0
+///    long double momentum = 1.25_TeV;              // == 1250.0
+///    long double lengthInUm = length / 1_um;       // == 1000000.0
+///    long double momentumInMeV = momentum / 1_MeV; // == 1250000.0
 ///
 /// .. warning::
 ///    Since using user-defined literals requires a namespace import of
@@ -94,8 +95,8 @@ namespace Acts {
 ///
 ///   .. code-block:: cpp
 ///
-///      double momentum = 100.0; // momentum is stored as native unit GeV
-///      double momentumInMeV = 10.0; // would be 0.01 in native units
+///      long double momentum = 100.0; // momentum is stored as native unit GeV
+///      long double momentumInMeV = 10.0; // would be 0.01 in native units
 ///
 /// - All input values must be given as ``numerical_value * unit_constant`` or
 ///   equivalently using the unit literals as ``value_unit``. The resulting
@@ -114,22 +115,22 @@ namespace Acts {
 ///    using namespace Acts::UnitLiterals;
 ///
 ///    // define input values w/ units (via unit constants)
-///    double width    = 12 * Acts::UnitConstants::mm;
-///    double mmuon    = 105.7 * Acts::UnitConstants::MeV;
+///    long double width    = 12 * Acts::UnitConstants::mm;
+///    long double mmuon    = 105.7 * Acts::UnitConstants::MeV;
 ///    // define input values w/ units (via unit user literals)
-///    double length   = 23_cm;
-///    double time     = 1214.2_ns;
-///    double angle    = 123_degree;
-///    double momentum = 2.5_TeV;
-///    double mass     = 511_keV;
-///    double velocity = 345_m / 1_s;
-///    double bfield   = 3.9_T;
+///    long double length   = 23_cm;
+///    long double time     = 1214.2_ns;
+///    long double angle    = 123_degree;
+///    long double momentum = 2.5_TeV;
+///    long double mass     = 511_keV;
+///    long double velocity = 345_m / 1_s;
+///    long double bfield   = 3.9_T;
 ///
 ///    // convert output values (via unit constants)
-///    double t_in_ns    = trackPars.time() / Acts::UnitConstants::ns;
+///    long double t_in_ns    = trackPars.time() / Acts::UnitConstants::ns;
 ///    // convert output values (via unit user literals)
-///    double x_in_mm   = trackPars.position()[ePos0] / 1_mm;
-///    double p_in_TeV = trackPars.absoluteMomentum() / 1_TeV;
+///    long double x_in_mm   = trackPars.position()[ePos0] / 1_mm;
+///    long double p_in_TeV = trackPars.absoluteMomentum() / 1_TeV;
 ///
 /// @endverbatim
 
@@ -139,74 +140,74 @@ namespace Acts {
 
 namespace UnitConstants {
 // Length, native unit mm
-constexpr double mm = 1.0;
-constexpr double fm = 1e-12 * mm;
-constexpr double pm = 1e-9 * mm;
-constexpr double nm = 1e-6 * mm;
-constexpr double um = 1e-3 * mm;
-constexpr double cm = 1e1 * mm;
-constexpr double m = 1e3 * mm;
-constexpr double km = 1e6 * mm;
+constexpr long double mm = 1.0;
+constexpr long double fm = 1e-12 * mm;
+constexpr long double pm = 1e-9 * mm;
+constexpr long double nm = 1e-6 * mm;
+constexpr long double um = 1e-3 * mm;
+constexpr long double cm = 1e1 * mm;
+constexpr long double m = 1e3 * mm;
+constexpr long double km = 1e6 * mm;
 // Shortcuts for commonly used area and volume units. This intentionally
 // contains not all possible combinations to avoid cluttering the namespace.
 // Missing area or volume units can always be defined on the fly using the
 // existing length units e.g. 1fm³ -> 1fm * 1fm * 1fm
 // Area, native unit mm²
-constexpr double mm2 = mm * mm;
-constexpr double cm2 = cm * cm;
-constexpr double m2 = m * m;
+constexpr long double mm2 = mm * mm;
+constexpr long double cm2 = cm * cm;
+constexpr long double m2 = m * m;
 // Volume, native unit mm³
-constexpr double mm3 = mm * mm * mm;
-constexpr double cm3 = cm * cm * cm;
-constexpr double m3 = m * m * m;
+constexpr long double mm3 = mm * mm * mm;
+constexpr long double cm3 = cm * cm * cm;
+constexpr long double m3 = m * m * m;
 // Time, native unit mm = [speed-of-light * time] = mm/s * s
 /// @note Depends on speed of light in SI units
-constexpr double s = 299792458000.0;  // = 299792458.0 * (m / 1.0) * 1.0;
-constexpr double fs = 1e-15 * s;
-constexpr double ps = 1e-12 * s;
-constexpr double ns = 1e-9 * s;
-constexpr double us = 1e-6 * s;
-constexpr double ms = 1e-3 * s;
-constexpr double min = 60.0 * s;
-constexpr double h = 3600.0 * s;
+constexpr long double s = 299792458000.0;  // = 299792458.0 * (m / 1.0) * 1.0;
+constexpr long double fs = 1e-15 * s;
+constexpr long double ps = 1e-12 * s;
+constexpr long double ns = 1e-9 * s;
+constexpr long double us = 1e-6 * s;
+constexpr long double ms = 1e-3 * s;
+constexpr long double min = 60.0 * s;
+constexpr long double h = 3600.0 * s;
 // Angles, native unit radian
-constexpr double mrad = 1e-3;
-constexpr double rad = 1.0;
-constexpr double degree = std::numbers::pi / 180. / rad;
+constexpr long double mrad = 1e-3;
+constexpr long double rad = 1.0;
+constexpr long double degree = std::numbers::pi / 180. / rad;
 // Energy/mass/momentum, native unit GeV
-constexpr double GeV = 1.0;
-constexpr double eV = 1e-9 * GeV;
-constexpr double keV = 1e-6 * GeV;
-constexpr double MeV = 1e-3 * GeV;
-constexpr double TeV = 1e3 * GeV;
-constexpr double J = 6241509074.460763 * GeV;
+constexpr long double GeV = 1.0;
+constexpr long double eV = 1e-9 * GeV;
+constexpr long double keV = 1e-6 * GeV;
+constexpr long double MeV = 1e-3 * GeV;
+constexpr long double TeV = 1e3 * GeV;
+constexpr long double J = 6241509074.460763 * GeV;
 /// atomic mass unit u
-constexpr double u = 0.93149410242;
+constexpr long double u = 0.93149410242;
 //     1eV/c² == 1.782662e-36kg
 //    1GeV/c² == 1.782662e-27kg
 // ->     1kg == (1/1.782662e-27)GeV/c²
 // ->      1g == (1/(1e3*1.782662e-27))GeV/c²
-constexpr double g = 1.0 / 1.782662e-24;
-constexpr double kg = 1.0 / 1.782662e-27;
+constexpr long double g = 1.0 / 1.782662e-24;
+constexpr long double kg = 1.0 / 1.782662e-27;
 /// Charge, native unit e (elementary charge)
-constexpr double e = 1.0;
+constexpr long double e = 1.0;
 /// Magnetic field, native unit (eV*s)/(e*m²)
 /// @note Depends on speed of light in SI units
-constexpr double T = 0.000299792458;  // = eV * s / (e * m2);
-constexpr double Gauss = 1e-4 * T;
-constexpr double kGauss = 1e-1 * T;
+constexpr long double T = 0.000299792458;  // = eV * s / (e * m2);
+constexpr long double Gauss = 1e-4 * T;
+constexpr long double kGauss = 1e-1 * T;
 /// Amount of substance, native unit mol
-constexpr double mol = 1.0;
+constexpr long double mol = 1.0;
 }  // namespace UnitConstants
 
 namespace UnitLiterals {
 // define user literal functions for the given unit constant
-#define ACTS_DEFINE_UNIT_LITERAL(name)                       \
-  constexpr double operator""_##name(long double x) {        \
-    return ::Acts::UnitConstants::name * x;                  \
-  }                                                          \
-  constexpr double operator""_##name(unsigned long long x) { \
-    return ::Acts::UnitConstants::name * x;                  \
+#define ACTS_DEFINE_UNIT_LITERAL(name)                            \
+  constexpr long double operator""_##name(long long double x) {   \
+    return ::Acts::UnitConstants::name * x;                       \
+  }                                                               \
+  constexpr long double operator""_##name(unsigned long long x) { \
+    return ::Acts::UnitConstants::name * x;                       \
   }
 ACTS_DEFINE_UNIT_LITERAL(fm)
 ACTS_DEFINE_UNIT_LITERAL(pm)
@@ -256,11 +257,11 @@ ACTS_DEFINE_UNIT_LITERAL(mol)
 /// Unit constants are intentionally not listed.
 namespace PhysicalConstants {
 // Speed of light
-constexpr double c = 1.0;
+constexpr long double c = 1.0;
 /// Reduced Planck constant h/2*pi.
 ///
-/// Computed from CODATA 2018 constants to double precision.
-constexpr double hbar =
+/// Computed from CODATA 2018 constants to long double precision.
+constexpr long double hbar =
     6.582119569509066e-25 * UnitConstants::GeV * UnitConstants::s;
 }  // namespace PhysicalConstants
