+++ Core/include/Acts/Detector/detail/PortalHelper.hpp
@@ -34,7 +34,7 @@ class Portal;
 /// gathered from the sub volumes, the binning description
 using PortalReplacement =
     std::tuple<std::shared_ptr<Experimental::Portal>, unsigned int, Direction,
-               std::vector<double>, AxisDirection>;
+               std::vector<long double>, AxisDirection>;
 
 namespace detail::PortalHelper {
 
@@ -62,7 +62,7 @@ void attachExternalNavigationDelegate(
 void attachDetectorVolumesUpdater(
     const GeometryContext& gctx, Portal& portal,
     const std::vector<std::shared_ptr<DetectorVolume>>& volumes,
-    const Direction& direction, const std::vector<double>& boundaries,
+    const Direction& direction, const std::vector<long double>& boundaries,
     const AxisDirection& binning);
 
 /// @brief Create and attach the multi link updator, the portal will get
