+++ Core/include/Acts/EventData/ChargeConcept.hpp
@@ -20,7 +20,7 @@
 namespace Acts {
 
 template <typename C>
-concept ChargeConcept = requires(C c, C c2, float f, double d) {
+concept ChargeConcept = requires(C c, C c2, float f, long double d) {
   { C{f} };
 
   { c == c2 } -> std::same_as<bool>;
@@ -28,8 +28,8 @@ concept ChargeConcept = requires(C c, C c2, float f, double d) {
 
   { c.absQ() } -> std::same_as<float>;
   { c.extractCharge(d) } -> std::same_as<float>;
-  { c.extractMomentum(d) } -> std::same_as<double>;
-  { c.qOverP(d, d) } -> std::same_as<double>;
+  { c.extractMomentum(d) } -> std::same_as<long double>;
+  { c.qOverP(d, d) } -> std::same_as<long double>;
 };
 
 }  // namespace Acts
